update of single inst files

+++ 0-setup.sh
@@ -29,9 +29,11 @@ fi
 echo "-------------------------------------------------"
 echo "Setting up mirrors for optimal download"
 echo "-------------------------------------------------"
-pacman -S --noconfirm pacman-contrib curl
-mv /etc/pacman.d/mirrorlist /etc/pacman.d/mirrorlist.backup
-curl -s "https://www.archlinux.org/mirrorlist/?country=DE&protocol=https&use_mirror_status=on" | sed -e 's/^#Server/Server/' -e '/^#/d' | rankmirrors -n 5 - > /etc/pacman.d/mirrorlist
+pacman -S --noconfirm pacman-contrib reflector
+if [[ -f /etc/pacman.d/mirrorlist]]; then
+    mv /etc/pacman.d/mirrorlist /etc/pacman.d/mirrorlist.backup
+fi
+reflector --country Germany --latest 10 --sort rate --save /etc/pacman.d/mirrorlist
 
 nc=$(grep -c ^processor /proc/cpuinfo)
 echo "You have " $nc" cores."
